% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_classe_pdm.R
\name{consulta_classe_pdm}
\alias{consulta_classe_pdm}
\title{Retorna os PDM de uma determinada classe}
\usage{
consulta_classe_pdm(classe)
}
\arguments{
\item{classe}{- O número da classe}
}
\value{
Um Dataframe com todos os PDM dentro de uma classe
}
\description{
Retorna os PDM de uma determinada classe
}
