% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_grupo_classe.R
\name{consulta_grupo_classe}
\alias{consulta_grupo_classe}
\title{Retornando as classes para um determinado grupo}
\usage{
consulta_grupo_classe(grupo)
}
\arguments{
\item{grupo}{- codigo de grupo de produtos}
}
\value{
Dataframe com a lista de classes dentro de um determinado produto
}
\description{
Retornando as classes para um determinado grupo
}
\examples{

}
