% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_material_licitacoes.R
\name{consulta_material_licitacoes}
\alias{consulta_material_licitacoes}
\title{Title}
\usage{
consulta_material_licitacoes(material)
}
\arguments{
\item{material}{- codigo do material}
}
\value{
Dataframe com as Licitacoes onde pelo menos um lote era desse produto
}
\description{
Title
}
