% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_pdm_materiais.R
\name{consulta_pdm_materiais}
\alias{consulta_pdm_materiais}
\title{Title}
\usage{
consulta_pdm_materiais(pdm)
}
\arguments{
\item{pdm}{- Código PDM}
}
\value{
- Materiais no PDM escolhido
}
\description{
Title
}
